LayoutTests/platform/mac/media/encrypted-media/videoCapabilities-contentType-parsing.html

<!doctype html>
<title>videoCapabilities contentType parsing</title>
<script src="../../../../media/video-test.js"></script>
<script>
window.addEventListener('load', async event => {
    const tests = [
        {
            "input": 'video/mp4XXX; codecs="avc1.64001F"',
            "success": false
        },
        {
            "input": '\nvideo/mp4\n; codecs="avc1.64001F"',
            "success": true
        }
    ];

    try {
        for (const { input, success } of tests) {
            consoleWrite("Test contentType " + input);
            const access = navigator.requestMediaKeySystemAccess('org.w3.clearkey', [{
                initDataTypes: ["cenc"],
                videoCapabilities: [{contentType : input}]
            }]);
            await (success ? shouldResolve(access) : shouldReject(access));
        }
    } catch(e) {
    } finally {
        endTest();
    }
});
</script>
